backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pillow>=10.0
python-multipart>=0.0.9
opencv-python-headless>=4.9.0
numpy>=1.26
